+++ test/valid-file-path.test.js
@@ -4,7 +4,7 @@ describe('Linux - Correct', () => {
   const correct = [
     'a/my_images/IMG--__23~123.min.jpeg',
     'modules/admin/foo_bar/test--__23~123.html.liquid',
-    'modules/admin cms/foo bar/test (Screenshot).liquid',
+    'modules/admin cms/foo bar/test (Screenshot)%20@.liquid',
     'a/foo/test+99.zip'
   ];
 
@@ -21,7 +21,6 @@ describe('Linux - Incorrect', () => {
     'a/foo/test".zip',
     'a/foo/test*.zip',
     'a/foo/test#.zip',
-    'a/foo/test@.zip',
     'a/foo/test<.zip',
     'a/foo/test|.zip',
     'a/foo/test?.zip'
@@ -38,7 +37,7 @@ describe('Windows - Correct', () => {
   const correct = [
     'a\\my_images\\IMG--__23~123.min.jpeg',
     'modules\\admin\\foo_bar\\test--__23~123.html.liquid',
-    'modules\\admin cms\\foo bar\\test (Screenshot).liquid',
+    'modules\\admin cms\\foo bar\\test (Screenshot)%20@.liquid',
     'a\\foo\\test+99.zip'
   ];
   correct.forEach(p => {
@@ -54,7 +53,6 @@ describe('Windows - Incorrect', () => {
     'a\\foo\\test".zip',
     'a\\foo\\test*.zip',
     'a\\foo\\test#.zip',
-    'a\\foo\\test@.zip',
     'a\\foo\\test<.zip',
     'a\\foo\\test|.zip',
     'a\\foo\\test?.zip'